requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pypdf>=3.0.0
python-dotenv>=1.0.0
openai>=1.0.0
aiofiles>=23.0.0